Add PostgreSQL backend support (#316)
* Add PostgreSQL backend support

garmin_connect_config_manager.get_db_params builds a DbParams for
db.type='postgresql', reading user/password/host/port/database from
the JSON config. GarminConnectConfig.json.example shows the new keys.

Also fix two pre-existing sqlite-only constructs surfaced by the
postgres backend:

- activities_db._create_sport_view / _create_course_view: use SQL '='
  and single-quoted literals (sqlite tolerates '==' and double-quoted
  strings; postgres reads "foo" as an identifier).
- summarydb.summary_base.__create_weeks_months_years_selectable: the
  per-period view selected sweat_loss_avg twice. Postgres rejects
  duplicate output columns. Second slot was meant to be sweat_loss.

* Fix import-side type issues exposed by postgres backend

Four bugs surfaced when running `--import` against the new postgres
backend. Sqlite was permissive enough to mask each one; postgres is
not. Fixed in place — sqlite and mysql users see no behaviour change.

- Device.serial_number is declared `Integer` but Garmin's Descent Mk3i
  family reports serials > 2**31 (e.g. 3476167836). The hard-coded
  fallback `unknown_device_serial_number = 9999999999` would also
  overflow on a strict 32-bit backend. Switch to BigInteger on the
  primary key and on the matching FK columns in DeviceInfo / File and
  on activities.device_serial_number / activity_laps.device_serial_number.

- monitoring_info.file_id was `Integer`, but the value passed in is the
  string id from `File.id` (which itself is `String`). Sqlite quietly
  coerced "269351985837" to a numeric value; postgres rejects the
  string-to-int4 cast outright. Switch monitoring_info.file_id to
  `String` so the FK shape matches `File.id`.

- monitoring_fit_file_processor.__unpack_tuple only unpacked when
  type(value) is tuple, but fitfile yields per-activity-type arrays as
  plain `list`. The list bled through to a `Float` column and the
  failure poisoned the SQLAlchemy session for every subsequent insert
  in the same file. Accept both list and tuple via isinstance().

* Plug remaining import-side gaps + include today in --latest

Three small bugs hit while running --import + --download --latest
against the postgres backend; bundled here because they share the same
"sqlite tolerated, postgres does not" pattern.

- scripts/garmindb_cli.py __get_date_and_days set days = (today - date).days,
  which omits today. With a one-day-ago anchor that pulled yesterday and
  the day before but silently skipped today. Bumping both --latest and
  --all branches by 1 includes today; when upstream has nothing yet the
  API returns empty and the importer is a no-op for that day.

- garmindb/fit_file_processor._write_device_info_entry treated a serial
  of None as "unset" (and synthesised an id from parent + device_type)
  but kept the literal 0 that fitfile emits for an invalid/unset serial.
  Multiple local sensors per FIT then collide on a shared 0 PK, the
  failure poisons the session, and the rest of the file is dropped.
  Treat 0 the same as None on SourceType.local.

- garmindb/garmindb/garmin_db.Device.local_device_serial_number returned
  a string ('%s%06d' % ...); the value lands in an Integer column. Wrap
  in int() so postgres does not have to coerce.

* Update garmindb_cli.py

---------

+++ garmindb/garmindb/garmin_db.py
@@ -8,7 +8,7 @@ import os
 import datetime
 import logging
 import re
-from sqlalchemy import Column, Integer, DateTime, Time, Float, String, Enum, ForeignKey, func, PrimaryKeyConstraint
+from sqlalchemy import Column, Integer, BigInteger, DateTime, Time, Float, String, Enum, ForeignKey, func, PrimaryKeyConstraint
 from sqlalchemy.ext.hybrid import hybrid_property
 
 import fitfile
@@ -59,7 +59,7 @@ class Device(GarminDb.Base, idbutils.DbObject):
 
     Manufacturer = idbutils.derived_enum.derive('Manufacturer', fitfile.Manufacturer, {'Microsoft' : 100001, 'Unknown': 100000})
 
-    serial_number = Column(Integer, primary_key=True)
+    serial_number = Column(BigInteger, primary_key=True)
     timestamp = Column(DateTime)
     device_type = Column(String)
     manufacturer = Column(Enum(Manufacturer))
@@ -74,7 +74,9 @@ class Device(GarminDb.Base, idbutils.DbObject):
     @classmethod
     def local_device_serial_number(cls, serial_number, device_type):
         """Return a synthetic serial number for a sub device composed of the parent's serial number and the sub device type."""
-        return '%s%06d' % (serial_number, device_type.value)
+        # The returned value lands in an Integer column. Build it as int so postgres
+        # (and any other strict-typed backend) does not have to coerce from a string.
+        return int('%s%06d' % (serial_number, device_type.value))
 
 
 class DeviceInfo(GarminDb.Base, idbutils.DbObject):
@@ -88,7 +90,7 @@ class DeviceInfo(GarminDb.Base, idbutils.DbObject):
 
     timestamp = Column(DateTime, nullable=False)
     file_id = Column(String, ForeignKey('files.id'))
-    serial_number = Column(Integer, ForeignKey('devices.serial_number'), nullable=False)
+    serial_number = Column(BigInteger, ForeignKey('devices.serial_number'), nullable=False)
     software_version = Column(String)
     cum_operating_time = Column(Time, nullable=False, default=datetime.time.min)
     battery_status = Column(Enum(fitfile.field_enums.BatteryStatus))
@@ -135,7 +137,7 @@ class File(GarminDb.Base, idbutils.DbObject):
     id = Column(String, primary_key=True)
     name = Column(String, unique=True)
     type = Column(Enum(FileType), nullable=False)
-    serial_number = Column(Integer, ForeignKey('devices.serial_number'))
+    serial_number = Column(BigInteger, ForeignKey('devices.serial_number'))
 
     @classmethod
     def s_get_id(cls, session, pathname):
